Fix UT done common checker

+++ EmrCloudApi/Requests/MedicalExamination/GetValidGairaiRihaRequest.cs
@@ -2,7 +2,7 @@
 {
     public class GetValidGairaiRihaRequest
     {
-        public long PtId { get; set; }
+        public int PtId { get; set; }
         public long RaiinNo { get; set; }
         public int SinDate { get; set; }
         public int SyosaiKbn { get; set; }
